fix mot so bugs nho nho, rat la met moi

+++ Grid.cpp
@@ -6,8 +6,8 @@
 
 Grid::Grid(){
 
-    for (size_t i=0; i<ROWS+HIDDEN_ROWS; i++){
-        for (size_t j=0; j<COLS; j++){
+    for (int i=0; i<ROWS+HIDDEN_ROWS; i++){
+        for (int j=0; j<COLS; j++){
             matrix[i][j] = {j, i-HIDDEN_ROWS, backgroundColor};
         }
     }
@@ -16,17 +16,17 @@ Grid::~Grid(){
 
 }
 void Grid::render(SDL_Renderer *renderer){
-    for (size_t i = 0; i <= ROWS; i++){
+    for (int i = 0; i <= ROWS; i++){
         SDL_SetRenderDrawColor(renderer, 100, 100, 0, 0);
         SDL_RenderDrawLine(renderer, xPos, i*TILE_SIZE + yPos, xPos + TILE_SIZE*COLS, i*TILE_SIZE + yPos);
     }
-    for (size_t i = 0; i <= COLS; i++){
+    for (int i = 0; i <= COLS; i++){
         SDL_SetRenderDrawColor(renderer, 100, 100, 0, 0);
         SDL_RenderDrawLine(renderer, i*TILE_SIZE + xPos, yPos, i*TILE_SIZE + xPos, yPos + TILE_SIZE*ROWS);
     }
 
-    for (size_t i = 0; i < ROWS; i++){
-        for (size_t j = 0; j < COLS; j++){
+    for (int i = 0; i < ROWS; i++){
+        for (int j = 0; j < COLS; j++){
             if (matrix[i+HIDDEN_ROWS][j].exist){
                 matrix[i+HIDDEN_ROWS][j].render(renderer);
             }
